Lot 3.1 : les quantités cessent d'être entières
quantite_disponible était un integer. Le référentiel livre pourtant des
kilos, des litres et des mètres carrés : 12,5 kg de cacao entraient en
stock pour 12, sans erreur ni trace. Au bout d'un an de réceptions,
l'écart entre le stock théorique et le comptage physique n'avait plus
d'explication.

La correction porte sur toute la chaîne, pas seulement sur la fiche de
stock — une quantité tronquée sur la ligne de vente l'est déjà avant
d'atteindre le stock.

  Base         six tables en decimal(15,3) : stocks, mouvements_stock,
               vente_details, achat_details, produits, transferts_stock
  Modèles      casts en float ; Produit::stockActuel() et ses voisines
               retypées int → float
  Contrôleurs  quatre intval / (int) qui tronquaient à la lecture ou à
               l'écriture
  Validation   App\Modules\Admin\Regles\Quantite, énoncée une fois pour
               onze contrôleurs qui la recopiaient — sept disaient
               integer, quatre numeric sans borner les décimales
  Navigateur   parseInt → quantiteSaisie(), step="0.001" sur onze champs

Trois décimales, et pas davantage : le gramme sur le kilo, le millilitre
sur le litre. Deux colonnes gardent leurs quatre décimales,
fiche_technique_details.quantite et ordres_production.quantite_cible :
la première n'est pas une quantité mais un coefficient de nomenclature —
0,0125 kg de colorant par unité — et arrondir un coefficient fausse
toute la série qu'il multiplie.

Quantite::physique() refuse le zéro, le négatif et la quatrième
décimale. Refuser plutôt qu'arrondir : la colonne n'en garde que trois,
et un arrondi silencieux est précisément ce que ce lot corrige. Le sens
d'un mouvement est porté par son type, jamais par le signe.

Un défaut de saisie trouvé au passage : les champs de quantité des
avoirs portaient min="0.01" sans step. En HTML la base du pas est min :
les seules valeurs acceptées étaient 0,01 / 1,01 / 2,01… et « 12 »
était refusé par le navigateur, sans message compréhensible.

Un point signalé et laissé au propriétaire : FneService transmet
intval($detail->quantite) à six endroits. Le code est gelé et aucune des
trois exceptions ne s'applique — le journal de bord dit ce qu'il faut
vérifier avant de décider.

186 tests, 2 213 assertions.

+++ app/Modules/Admin/Controleurs/BonLivraisonControleur.php
@@ -72,8 +72,10 @@ class BonLivraisonControleur extends Controller
                 ->where('point_de_vente_id', $pointDeVenteId)
                 ->first();
 
-            $stockDispo    = $stock ? max(0, (int) $stock->quantite_disponible) : 0;
-            $qteCom        = (int) $detail->quantite;
+            // En float : le stock se compte aussi en kilos et en litres, et
+            // (int) faisait apparaitre 12,5 kg comme 12 disponibles.
+            $stockDispo    = $stock ? max(0, (float) $stock->quantite_disponible) : 0;
+            $qteCom        = (float) $detail->quantite;
             $qteSuggestion = min($qteCom, $stockDispo);
             $estInsuffisant = $stockDispo < $qteCom;
